+++ R/stats.R
@@ -67,15 +67,13 @@ test_as_row <- function(
 #' details and examples for more.
 #' @return Validated list of instructions that may be used in the
 #' \code{\link{run_algos}} function.
-#' @details Valid list names (algorithm function names in \code{mdsstat}) are
-#' currently \code{\link{shewhart}}, \code{\link{prr}}, and
-#' \code{\link{poisson_rare}}. Each algorithm may be named multiple times (to
+#' @details Each algorithm may be named multiple times (to
 #' allow running of multiple parameter settings). Do not specify the \code{df}
 #' parameter.
 #' @examples
 #' x <- list(prr=list(),
-#'   shewhart=list(),
-#'   shewhart=list(ts_event=c(Rate="rate"), we_rule=2L),
+#'   xbar=list(),
+#'   xbar=list(ts_event=c(Rate="rate"), we_rule=2),
 #'   poisson_rare=list(p_rate=0.3))
 #' define_algos(x)
 #' @export
@@ -88,7 +86,7 @@ define_algos <- function(
   algolist <- c("poisson_rare", "prr", "shewhart", "xbar",
                 "cusum", "ror",
                 "sprt", "gps", "bcpnn",
-                "ewma")
+                "ewma", "cp_mean")
   # algolist <- ls("package:mdsstat")[grepl("\\.mds_ts$", ls("package:mdsstat"))]
   # algolist <- gsub("\\.mds_ts$", "", algolist)
   if (!all(names(algos) %in% algolist)){
@@ -159,8 +157,8 @@ define_algos <- function(
 #' data <- mds_ts[[1]]
 #' data$rate <- data$nA / data$exposure
 #' x <- list(prr=list(),
-#'   shewhart=list(),
-#'   shewhart=list(ts_event=c(Rate="rate"), we_rule=2L),
+#'   xbar=list(),
+#'   xbar=list(ts_event=c(Rate="rate"), we_rule=2),
 #'   poisson_rare=list(p_rate=0.3))
 #' algos <- define_algos(x)
 #' run_algos(data, algos)
@@ -221,9 +219,7 @@ run_algos.default <- function(
                "skip, stop, warn"))
   }
   # Define DPA algorithms currently in mdsstat
-  dpaalgos <- c("poisson_rare", "prr", "shewhart",
-                "cusum", "ror",
-                "sprt", "gps", "bcpnn")
+  dpaalgos <- c("prr", "ror", "gps", "bcpnn")
 
   if (dataframe){
     stats <- data.frame()